+++ examples/apps/todomvc-rt/src/footer/footer.test.js
@@ -1,25 +1,22 @@
 import { it, expect } from "vitest"
-import { createStore } from "@inglorious/store"
+import { trigger } from "@inglorious/store/test"
 
 import { footer } from "./footer"
 
 it("should update the filter on filterClick", () => {
-  const stateBefore = {
-    footer: { id: "footer", type: "footer", activeFilter: "all" },
+  const entityBefore = {
+    id: "footer",
+    type: "footer",
+    activeFilter: "all",
   }
   const event = { type: "filterClick", payload: "completed" }
-  const stateAfter = {
-    footer: { id: "footer", type: "footer", activeFilter: "completed" },
+  const entityAfter = {
+    id: "footer",
+    type: "footer",
+    activeFilter: "completed",
   }
 
-  doTest(stateBefore, event, stateAfter)
+  const { entity } = trigger(entityBefore, footer[event.type], event.payload)
+
+  expect(entity).toEqual(entityAfter)
 })
-
-function doTest(stateBefore, event, stateAfter) {
-  const store = createStore({ types: { footer }, entities: stateBefore })
-
-  store.dispatch(event)
-  const state = store.getState()
-
-  expect(state).toEqual(stateAfter)
-}